Avoid getting detected as a spambot.
Some servers (e.g., the servers on the AccessIRC network at time of
writing) refuse connections if the unused "hostname" and "servername" in
the USER line are given as "x". This patch uses the username as the
hostname and servername, in order to avoid the problem.

darcs-hash:20051119225814-c8bf2-2736e83bca36f5e38044c4176d3496ca4848e06d